refactor(validate): pull out the match functions
