+++ src/validate/index.js
@@ -1,100 +1,7 @@
 "use strict";
 
-function propertyEquals(pTo, pRule, pProperty) {
-    return pRule.to.hasOwnProperty(pProperty)
-        ? pTo[pProperty] === pRule.to[pProperty]
-        : true;
-}
-
-function intersects(pToDependencyTypes, pRuleDependencyTypes) {
-    return pToDependencyTypes.some(
-        pDepType => pRuleDependencyTypes.some(
-            pRDepType => pDepType === pRDepType
-        )
-    );
-}
-
-
-/* if there is at least one group expression in the given pRulePath
-   return the first matched one.
-   return null in all other cases
-
-   This fills our current need. Later we can expand it to return all group
-   matches.
-*/
-
-function extractGroups(pRule, pActualPath) {
-    let lRetval = [];
-
-    if (Boolean(pRule.path)) {
-        let lMatchResult = pActualPath.match(pRule.path);
-
-        if (Boolean(lMatchResult) && lMatchResult.length > 1) {
-            lRetval = lMatchResult;
-        }
-    }
-    return lRetval;
-}
-
-function replaceGroupPlaceholders(pString, pExtractedGroups) {
-    return pExtractedGroups.reduce(
-        (pAll, pThis, pIndex) => pAll.replace(`$${pIndex}`, pThis),
-        pString
-    );
-}
-
-function isModuleOnlyRule(pRule){
-    return pRule.from.hasOwnProperty("orphan");
-}
-
-function matchModuleRule(pModule){
-    return pRule => (
-        pModule.orphan === true &&
-            pRule.from.orphan === true
-    ) && (!pRule.from.path ||
-                pModule.source.match(pRule.from.path)
-    ) && (!pRule.from.pathNot ||
-            !(pModule.source.match(pRule.from.pathNot))
-    );
-}
-matchModuleRule.isInteresting = pRule => isModuleOnlyRule(pRule);
-
-function matchDependencyRule(pFrom, pTo) {
-    return pRule => {
-        const lGroups = extractGroups(pRule.from, pFrom.source);
-
-        /*
-         * the replace("$1", lGroup) things below are a bit simplistic (they
-         * also match \$, which they probably shouldn't) - but good enough for
-         * now.
-         */
-        return (!pRule.from.path ||
-                pFrom.source.match(pRule.from.path)
-        ) && (!pRule.from.pathNot ||
-                !(pFrom.source.match(pRule.from.pathNot))
-        ) && (!pRule.to.path ||
-                (lGroups.length > 0
-                    ? pTo.resolved.match(replaceGroupPlaceholders(pRule.to.path, lGroups))
-                    : pTo.resolved.match(pRule.to.path)
-                )
-        ) && (!Boolean(pRule.to.pathNot) ||
-                !(lGroups.length > 0
-                    ? pTo.resolved.match(replaceGroupPlaceholders(pRule.to.pathNot, lGroups))
-                    : pTo.resolved.match(pRule.to.pathNot)
-                )
-        ) && (!pRule.to.dependencyTypes ||
-                intersects(pTo.dependencyTypes, pRule.to.dependencyTypes)
-        ) && (!pRule.to.hasOwnProperty("moreThanOneDependencyType") ||
-                pTo.dependencyTypes.length > 1
-        ) && (!pRule.to.license ||
-                pTo.license && pTo.license.match(pRule.to.license)
-        ) && (!pRule.to.licenseNot ||
-                pTo.license && !pTo.license.match(pRule.to.licenseNot)
-        ) && propertyEquals(pTo, pRule, "couldNotResolve") &&
-                 propertyEquals(pTo, pRule, "circular");
-    };
-}
-matchDependencyRule.isInteresting = pRule => !isModuleOnlyRule(pRule);
+const matchModuleRule     = require('./matchModuleRule');
+const matchDependencyRule = require('./matchDependencyRule');
 
 function compareSeverity(pFirst, pSecond) {
     const SEVERITY2INT = {
@@ -106,16 +13,16 @@ function compareSeverity(pFirst, pSecond) {
     return SEVERITY2INT[pFirst.severity] - SEVERITY2INT[pSecond.severity];
 }
 
-function validateAgainstRules(pRuleSet, pFrom, pTo, pMatchFunction) {
+function validateAgainstRules(pRuleSet, pFrom, pTo, pMatchModule) {
     let lFoundRuleViolations = [];
     let lRetval = {valid:true};
 
     pRuleSet.forbidden = pRuleSet.forbidden || [];
 
     if (pRuleSet.allowed){
-        const lInterestingAllowedRules = pRuleSet.allowed.filter(pMatchFunction.isInteresting);
+        const lInterestingAllowedRules = pRuleSet.allowed.filter(pMatchModule.isInteresting);
 
-        if (lInterestingAllowedRules.length > 0 && !(lInterestingAllowedRules.some(pMatchFunction(pFrom, pTo)))){
+        if (lInterestingAllowedRules.length > 0 && !(lInterestingAllowedRules.some(pMatchModule.match(pFrom, pTo)))){
             lFoundRuleViolations.push({
                 severity: pRuleSet.allowedSeverity,
                 name: "not-in-allowed"
@@ -127,8 +34,8 @@ function validateAgainstRules(pRuleSet, pFrom, pTo, pMatchFunction) {
         .concat(
             pRuleSet
                 .forbidden
-                .filter(pMatchFunction.isInteresting)
-                .filter(pMatchFunction(pFrom, pTo))
+                .filter(pMatchModule.isInteresting)
+                .filter(pMatchModule.match(pFrom, pTo))
                 .map(pMatchedRule => ({
                     severity : pMatchedRule.severity,
                     name     : pMatchedRule.name
@@ -169,7 +76,6 @@ module.exports = {
         if (!pValidate) {
             return {valid:true};
         }
-        // return {};
         return validateAgainstRules(pRuleSet, pModule, {}, matchModuleRule);
     },
     dependency: (pValidate, pRuleSet, pFrom, pTo) => {
@@ -184,4 +90,3 @@ module.exports = {
    - we only use it from within the module with two fixed values
    - the propertyEquals function is not exposed externaly
  */
-/* eslint security/detect-object-injection: 0, complexity: 0 */
